Minor markup changes
